Extract `.strings` validation helper in dedicated file
